+++ src/main/java/ch/heigvd/res/chill/domain/wasadigi/Duvel.java
@@ -1,13 +1,13 @@
-package ch.heigvd.res.chill.domain.AxelVallon;
+package ch.heigvd.res.chill.domain.wasadigi;
 
 import ch.heigvd.res.chill.domain.IProduct;
 
 import java.math.BigDecimal;
 
-public class Boxer implements IProduct {
+public class Duvel implements IProduct {
 
-  public final static String NAME = "Boxer";
-  public final static BigDecimal PRICE = new BigDecimal(2.0);
+  public final static String NAME = "Duvel";
+  public final static BigDecimal PRICE = new BigDecimal(3.8);
 
   @Override
   public String getName() {
