+++ src/main/java/ch/heigvd/res/chill/domain/wasadigi/LaCorne.java
@@ -1,13 +1,13 @@
-package ch.heigvd.res.chill.domain.AxelVallon;
+package ch.heigvd.res.chill.domain.wasadigi;
 
 import ch.heigvd.res.chill.domain.IProduct;
 
 import java.math.BigDecimal;
 
-public class Duvel implements IProduct {
+public class LaCorne implements IProduct {
 
-  public final static String NAME = "Duvel";
-  public final static BigDecimal PRICE = new BigDecimal(3.8);
+  public final static String NAME = "La Corne";
+  public final static BigDecimal PRICE = new BigDecimal(4.85);
 
   @Override
   public String getName() {
