+++ src/main/java/ch/heigvd/res/chill/domain/wasadigi/PunkIPA.java
@@ -1,13 +1,13 @@
-package ch.heigvd.res.chill.domain.AxelVallon;
+package ch.heigvd.res.chill.domain.wasadigi;
 
 import ch.heigvd.res.chill.domain.IProduct;
 
 import java.math.BigDecimal;
 
-public class LaFranche implements IProduct {
+public class PunkIPA implements IProduct {
 
-  public final static String NAME = "LaFranche";
-  public final static BigDecimal PRICE = new BigDecimal(4.8);
+  public final static String NAME = "Punk IPA";
+  public final static BigDecimal PRICE = new BigDecimal(4.2);
 
   @Override
   public String getName() {
