+++ src/main/java/ch/heigvd/res/chill/protocol/GreetingsRequest.java
@@ -4,7 +4,5 @@ import lombok.Data;
 
 @Data
 public class GreetingsRequest implements IMessage {
-
   private final String text;
-
 }
