+++ src/test/java/ch/heigvd/res/chill/domain/AxelVallon/TroisDameTest.java
@@ -1,5 +1,6 @@
 package ch.heigvd.res.chill.domain.AxelVallon;
 
+import ch.heigvd.res.chill.domain.AxelVallon.TroisDame;
 import ch.heigvd.res.chill.domain.Bartender;
 import ch.heigvd.res.chill.protocol.OrderRequest;
 import ch.heigvd.res.chill.protocol.OrderResponse;
@@ -9,22 +10,22 @@ import java.math.BigDecimal;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-class BoxerTest {
+class TroisDameTest {
 
   @Test
-  void thePriceAndNameForBoxerShouldBeCorrect() {
-    Boxer beer = new Boxer();
-    assertEquals(beer.getName(), Boxer.NAME);
-    assertEquals(beer.getPrice(), Boxer.PRICE);
+  void thePriceAndNameForPunkIPAShouldBeCorrect() {
+    TroisDame beer = new TroisDame();
+    assertEquals(beer.getName(), TroisDame.NAME);
+    assertEquals(beer.getPrice(), TroisDame.PRICE);
   }
 
   @Test
-  void aBartenderShouldAcceptAnOrderForBoxer() {
+  void aBartenderShouldAcceptAnOrderForPunkIPA() {
     Bartender jane = new Bartender();
-    String productName = "ch.heigvd.res.chill.domain.wasadigi.Boxer";
+    String productName = "ch.heigvd.res.chill.domain.AxelVallon.TroisDame";
     OrderRequest request = new OrderRequest(3, productName);
     OrderResponse response = jane.order(request);
-    BigDecimal expectedTotalPrice = Boxer.PRICE.multiply(new BigDecimal(3));
+    BigDecimal expectedTotalPrice = TroisDame.PRICE.multiply(new BigDecimal(3));
     assertEquals(expectedTotalPrice, response.getTotalPrice());
   }
