+++ src/test/java/ch/heigvd/res/chill/domain/wasadigi/BoxerTest.java
@@ -1,0 +1,31 @@
+package ch.heigvd.res.chill.domain.wasadigi;
+
+import ch.heigvd.res.chill.domain.Bartender;
+import ch.heigvd.res.chill.protocol.OrderRequest;
+import ch.heigvd.res.chill.protocol.OrderResponse;
+import org.junit.jupiter.api.Test;
+
+import java.math.BigDecimal;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
+class BoxerTest {
+
+  @Test
+  void thePriceAndNameForBoxerShouldBeCorrect() {
+    Boxer beer = new Boxer();
+    assertEquals(beer.getName(), Boxer.NAME);
+    assertEquals(beer.getPrice(), Boxer.PRICE);
+  }
+
+  @Test
+  void aBartenderShouldAcceptAnOrderForBoxer() {
+    Bartender jane = new Bartender();
+    String productName = "ch.heigvd.res.chill.domain.wasadigi.Boxer";
+    OrderRequest request = new OrderRequest(3, productName);
+    OrderResponse response = jane.order(request);
+    BigDecimal expectedTotalPrice = Boxer.PRICE.multiply(new BigDecimal(3));
+    assertEquals(expectedTotalPrice, response.getTotalPrice());
+  }
+
+}
