+++ src/test/java/ch/heigvd/res/chill/domain/wasadigi/DuvelTest.java
@@ -1,4 +1,4 @@
-package ch.heigvd.res.chill.domain.AxelVallon;
+package ch.heigvd.res.chill.domain.wasadigi;
 
 import ch.heigvd.res.chill.domain.Bartender;
 import ch.heigvd.res.chill.protocol.OrderRequest;
@@ -9,22 +9,22 @@ import java.math.BigDecimal;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-class LaCorneTest {
+class DuvelTest {
 
   @Test
-  void thePriceAndNameForLaCorneShouldBeCorrect() {
-    LaCorne beer = new LaCorne();
-    assertEquals(beer.getName(), LaCorne.NAME);
-    assertEquals(beer.getPrice(), LaCorne.PRICE);
+  void thePriceAndNameForPunkIPAShouldBeCorrect() {
+    Duvel beer = new Duvel();
+    assertEquals(beer.getName(), Duvel.NAME);
+    assertEquals(beer.getPrice(), Duvel.PRICE);
   }
 
   @Test
-  void aBartenderShouldAcceptAnOrderForLaFranche() {
+  void aBartenderShouldAcceptAnOrderForPunkIPA() {
     Bartender jane = new Bartender();
-    String productName = "ch.heigvd.res.chill.domain.wasadigi.LaCorne";
+    String productName = "ch.heigvd.res.chill.domain.wasadigi.Duvel";
     OrderRequest request = new OrderRequest(3, productName);
     OrderResponse response = jane.order(request);
-    BigDecimal expectedTotalPrice = LaCorne.PRICE.multiply(new BigDecimal(3));
+    BigDecimal expectedTotalPrice = Duvel.PRICE.multiply(new BigDecimal(3));
     assertEquals(expectedTotalPrice, response.getTotalPrice());
   }
