+++ src/test/java/ch/heigvd/res/chill/domain/wasadigi/LaCorneTest.java
@@ -1,4 +1,4 @@
-package ch.heigvd.res.chill.domain.AxelVallon;
+package ch.heigvd.res.chill.domain.wasadigi;
 
 import ch.heigvd.res.chill.domain.Bartender;
 import ch.heigvd.res.chill.protocol.OrderRequest;
@@ -9,22 +9,22 @@ import java.math.BigDecimal;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-class LaFrancheTest {
+class LaCorneTest {
 
   @Test
-  void thePriceAndNameForLaFrancheShouldBeCorrect() {
-    LaFranche beer = new LaFranche();
-    assertEquals(beer.getName(), LaFranche.NAME);
-    assertEquals(beer.getPrice(), LaFranche.PRICE);
+  void thePriceAndNameForLaCorneShouldBeCorrect() {
+    LaCorne beer = new LaCorne();
+    assertEquals(beer.getName(), LaCorne.NAME);
+    assertEquals(beer.getPrice(), LaCorne.PRICE);
   }
 
   @Test
   void aBartenderShouldAcceptAnOrderForLaFranche() {
     Bartender jane = new Bartender();
-    String productName = "ch.heigvd.res.chill.domain.wasadigi.LaFranche";
+    String productName = "ch.heigvd.res.chill.domain.wasadigi.LaCorne";
     OrderRequest request = new OrderRequest(3, productName);
     OrderResponse response = jane.order(request);
-    BigDecimal expectedTotalPrice = LaFranche.PRICE.multiply(new BigDecimal(3));
+    BigDecimal expectedTotalPrice = LaCorne.PRICE.multiply(new BigDecimal(3));
     assertEquals(expectedTotalPrice, response.getTotalPrice());
   }
